fix(prompt-kit): require P00 doctrine installation
Replace acknowledgment-only P00 behavior with a commit-required repository doctrine workflow, repair contradictory workbook routing metadata, and add fail-closed action-commitment contracts.
