+++ configs/prompt_kit/v39_core_prompt_action_overrides.json
@@ -1,0 +1,140 @@
+{
+  "schema_version": 1,
+  "policy": {
+    "description": "Prompts that claim setup, installation, build, repair, deployment, or execution must require the matching mutation and proof rather than permit acknowledgment-only completion.",
+    "action_claim_tokens": [
+      "install",
+      "setup",
+      "build",
+      "execute",
+      "repair",
+      "deploy",
+      "configure",
+      "upgrade",
+      "merge",
+      "release"
+    ],
+    "required_action_markers": [
+      "Modify tracked repository files",
+      "git diff --check",
+      "git commit -m",
+      "git push -u origin",
+      "Open or update a pull request",
+      "exact blocker"
+    ],
+    "forbidden_completion_shapes": [
+      "acknowledgment only",
+      "summary only",
+      "rewritten prompt only",
+      "plan only",
+      "handoff only",
+      "preflight only"
+    ]
+  },
+  "prompts": [
+    {
+      "prompt_id": "P00",
+      "surface_family": "standard_ai",
+      "execution_shape": "repo_mutation",
+      "seq": "00",
+      "prompt_type": "SETUP",
+      "prompt_class": "SETUP / DOCTRINE",
+      "sprint_path_role": "Install universal harness discipline as tracked repository doctrine",
+      "use_for_progress": "YES",
+      "prompt_name": "Commit-Required Harness Doctrine Installer",
+      "use_this_when": "Baseline harness doctrine must be installed or repaired in the repository before serious work.",
+      "inspect_first": "Current repo rules, AGENTS.md or equivalent instruction surfaces, validators, tests, manifests, branch state, and PR conventions.",
+      "expected_output": "Repo-native doctrine and an enforcement test or validator committed to the relevant repository, with push or PR evidence.",
+      "next_step": "Use the task-specific prompt only after the doctrine commit and PR update are proven.",
+      "proof_gate": "Changed tracked doctrine files, enforcement validation, commit SHA, branch state, push confirmation, and opened or updated PR evidence.",
+      "color": "Slate",
+      "lines": [
+        "INSTALL THE HARNESS DOCTRINE NOW. DO NOT MERELY ACKNOWLEDGE, SUMMARIZE, OR REPEAT THESE RULES.",
+        "",
+        "Repo: resolve the current repository from the environment when no explicit repo is supplied.",
+        "Branch or worktree: preserve existing work; use the current safe branch or create an isolated feature branch when needed.",
+        "Sprint: Commit-Required Harness Doctrine Installation.",
+        "Lane: repository doctrine and enforcement.",
+        "Owned scope: repo-native agent rules, doctrine references, focused validators or tests, and the PR metadata needed to prove installation.",
+        "Forbidden scope: unrelated product changes, destructive cleanup, force-push, default-branch mutation, secrets, private evidence, or generated runtime junk.",
+        "Expected artifacts: tracked doctrine, an enforcement test or validator, validation output, commit SHA, push confirmation, and an opened or updated PR.",
+        "Validation order: targeted doctrine contract, relevant instruction or harness validators, broader safe checks, then final Git review.",
+        "",
+        "START WITH BOUNDED PREFLIGHT",
+        "Run:",
+        "git status --short",
+        "git branch --show-current",
+        "git log --oneline --decorate -5",
+        "If the worktree is dirty and this sprint does not own those changes, preserve them and use an isolated worktree or branch.",
+        "",
+        "INSPECT BEFORE INVENTING",
+        "Inspect existing AGENTS.md, CLAUDE.md, repository instruction files, harness doctrine, codebase maps, workflow registries, validators, tests, hooks, generated-output policy, and branch or PR conventions.",
+        "Reuse the repository's established authority surface. Do not create a second competing doctrine file when a canonical one already exists.",
+        "",
+        "REQUIRED REPOSITORY ACTION",
+        "Modify tracked repository files so the doctrine is actually installed, not merely active in this conversation.",
+        "At minimum:",
+        "- write or repair the canonical repo agent-rule or doctrine surface",
+        "- preserve task-specific instructions as higher-priority execution contracts",
+        "- add or update a focused test, validator, hook, or manifest rule that rejects acknowledgment-only setup",
+        "- connect the doctrine to the repository's existing harness or documentation index when that pattern exists",
+        "",
+        "The installed doctrine must require agents to name:",
+        "- repo",
+        "- branch or worktree",
+        "- PR or sprint",
+        "- lane",
+        "- owned scope",
+        "- forbidden scope",
+        "- expected artifacts",
+        "- validation order when specified",
+        "",
+        "It must enforce this executable loop:",
+        "request -> evidence review -> bounded decision -> repo or Git or GitHub mutation -> artifacts -> validation -> report -> next decision",
+        "",
+        "It must also enforce:",
+        "- evidence before confidence",
+        "- preservation before destructive cleanup",
+        "- existing contracts and helpers before invention",
+        "- no completion claim without checks, artifacts, commit or GitHub evidence",
+        "- no substitution of a rewritten prompt, plan, or handoff for requested repository work",
+        "- task-specific execution rules override generic closeout behavior",
+        "",
+        "ACTION-COMMITMENT RULE",
+        "Any prompt that claims it will install, set up, build, execute, repair, configure, upgrade, deploy, merge, or release something must require the corresponding mutation and proof.",
+        "A title or expected output that claims action while the prompt permits acknowledgment, advice, or a plan is invalid.",
+        "",
+        "VALIDATE AND COMMIT",
+        "Run the strongest practical focused checks and any repository-owned instruction or harness validators.",
+        "Then run:",
+        "git diff --check",
+        "git status --short",
+        "git diff --stat",
+        "git diff",
+        "Stage only owned tracked files.",
+        "Run git commit -m with a useful doctrine-installation message.",
+        "Run git push -u origin for the safe feature branch.",
+        "Open or update a pull request and record the exact commit SHA and validation evidence.",
+        "",
+        "ALREADY-COMPLETE PATH",
+        "Do not create meaningless churn when the exact doctrine and enforcement already exist.",
+        "Instead, prove the tracked files and passing checks, then make an owned GitHub mutation such as correcting the PR body or recording the validation evidence.",
+        "",
+        "BLOCKER PATH",
+        "When repository mutation is genuinely impossible, stop with the exact blocker, the smallest patch or file content that should be applied, and one safest next command.",
+        "",
+        "INVALID RESPONSES",
+        "- acknowledgment only",
+        "- summary only",
+        "- rewritten prompt only",
+        "- plan only",
+        "- handoff only",
+        "- preflight only",
+        "- a request for permission when the bounded mutation is safe",
+        "",
+        "FINAL REPORT",
+        "Report repo, branch or worktree, PR or sprint, lane, owned and forbidden scope, files changed, validation commands and results, commit SHA, push or PR mutation, gaps, final Git state, and one exact next command."
+      ]
+    }
+  ]
+}
